+++ module/scripts/apps/crawl-tracker.mjs
@@ -1,9 +1,10 @@
 import actorCarousel from "./actor-carousel.mjs";
+import TimePasses from "/module/scripts/apps/functions/timepasses.mjs";
 
 const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
 
 export default class crawlTracker extends HandlebarsApplicationMixin(ApplicationV2) {
-
+    
     constructor() {
 		super();
         this._dragDrop = this.options.dragDrop.map(d => {
@@ -171,69 +172,8 @@ export default class crawlTracker extends HandlebarsApplicationMixin(Application
         this._addParty();
     }
 
-    static async timePasses(){
-        new foundry.applications.api.DialogV2({
-          window: { title: "Time Passes" },
-          content: `
-            <div style="padding: 10px;">
-              <label for="minutes" style="font-family: 'Montserrat'; font-size: 14px;">
-                Enter Minutes to Pass:
-              </label>
-              <input id="minutes" name="minutes" type="number" min="1" value="10" style="width: 100%; margin-top: 5px;">
-            </div>
-          `,
-          buttons: [{
-            action: "advance",
-            label: "Pass Time",
-            default: true,
-            callback: (event, button, dialog) => {
-              return button.form.elements.minutes.valueAsNumber;
-            }
-          }],
-          submit: async (result) => {
-            if (isNaN(result) || result <= 0) {
-              return ui.notifications.warn("Please enter a valid number of minutes greater than 0.");
-            }
-            
-            const seconds = result * 60;
-            await game.time.advance(seconds);
-            
-            const players = game.combat.combatants.filter(c => c.system.type === "Player");
-            let expiredEffectsList = [];
-            for (const combatant of players) {
-              const actor = game.actors.get(combatant.actorId);
-              if (!actor) continue;
-
-              const effectsToRemove = actor.items.filter(item => 
-                item.type === "Effect" &&
-                item.system.duration &&
-                item.system.duration.type === "rounds" &&
-                Number(item.system.duration.value) > 0
-              );
-              if (effectsToRemove.length > 0) {
-                const names = effectsToRemove.map(item => item.name);
-                const effectIds = effectsToRemove.map(item => item.id);
-                await actor.deleteEmbeddedDocuments("Item", effectIds);
-                expiredEffectsList.push(`<strong>${actor.name}</strong> - ${names.join(", ")}`);
-              }
-            }
-            const effectsExpiredText = expiredEffectsList.length > 0 ? expiredEffectsList.join("<br>") : "None";
-            
-            const timePassesCard = `
-              <div class="shadowdark">
-                <h2 class="centered" style="font-family: 'Montserrat-Medium';">Time Passes</h2>
-                <p><strong>Minutes Passed:</strong> ${result}</p>
-                <p><strong>Effects Expired:</strong><br>${effectsExpiredText}</p>
-              </div>
-            `;
-            await ChatMessage.create({ content: timePassesCard });
-            ui.notifications.info(`Time advanced by ${result} minute(s).`);
-            
-            await this._checkForEncounter(3);
-          },
-          rejectClose: false,
-          modal: true
-        }).render({ force: true });
+    static async timePasses() {
+        this._timePasses();
       }  
       
     static async moralCheck(mode="individual"){
