+++ module/templates/chats/timepasses-card.hbs
@@ -1,0 +1,5 @@
+<div class="shadowdark">
+  <h2 class="centered" style="font-family: 'Montserrat-Medium';">Time Passes</h2>
+  <p><strong>Minutes Passed:</strong> {{minutes}}</p>
+  <p><strong>Effects Expired:</strong><br>{{expiredEffectsText}}</p>
+</div>
