+++ module/templates/dialogs/timepasses-dialog.hbs
@@ -1,0 +1,6 @@
+<div style="padding: 10px;">
+  <label for="minutes" style="font-family: 'Montserrat'; font-size: 14px;">
+    Enter Minutes to Pass:
+  </label>
+  <input id="minutes" name="minutes" type="number" min="1" value="10" style="width: 100%; margin-top: 5px;">
+</div>
